requirements.txt
requests
pycryptodomex